refactor(db): unify repositories with single Querier pattern
Migrate all 22 repository packages from dual-querier branching to unified
r.Q() pattern, eliminating massive code duplication.

Changes:
- Add cmd/sqlc-gen/ tool combining normalization, type aliases, and querier generation
- Create internal/infrastructure/database/unified/ package with generated code
- Migrate all persistence repositories to use r.Q() instead of sqlite/postgres branching
- Simplify BaseRepository from 258 to 54 lines
- Add Q() method to TransactionContext for unified transaction support
- Remove deprecated QueryRouter, generic_mapper, and router files

Metrics:
- Branching: 162 → 13 instances (remaining are legitimate DB-specific SQL)
- Duplicate converters: ~200 → 0
- Net code reduction: ~7,800 lines

+++ internal/infrastructure/database/queries/postgres/enrichment_pipeline.sql
@@ -70,7 +70,7 @@ DELETE FROM enrichment_pipelines WHERE id = $1;
 DELETE FROM enrichment_pipelines WHERE media_type = $1;
 
 -- name: GetNextPipelinePosition :one
-SELECT COALESCE(MAX(position), 0) + 1 as next_position
+SELECT (COALESCE(MAX(position), 0) + 1)::bigint as next_position
 FROM enrichment_pipelines
 WHERE media_type = $1;
